[knx] Improve handling of serial gateways (#13897)
* [knx] Improve handling of serial gateways

Take over initialization logic from KNX IP gateway for serial gateway.
Properly re-initialize serial gateway after configuration changes done in UI.
Fixes #13892.

+++ bundles/org.openhab.binding.knx/src/main/java/org/openhab/binding/knx/internal/config/BridgeConfiguration.java
@@ -12,8 +12,6 @@
  */
 package org.openhab.binding.knx.internal.config;
 
-import java.math.BigDecimal;
-
 import org.eclipse.jdt.annotation.NonNullByDefault;
 
 /**
@@ -25,23 +23,23 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 @NonNullByDefault
 public class BridgeConfiguration {
     private int autoReconnectPeriod = 0;
-    private BigDecimal readingPause = BigDecimal.valueOf(0);
-    private BigDecimal readRetriesLimit = BigDecimal.valueOf(0);
-    private BigDecimal responseTimeout = BigDecimal.valueOf(0);
+    private int readingPause = 0;
+    private int readRetriesLimit = 0;
+    private int responseTimeout = 0;
 
     public int getAutoReconnectPeriod() {
         return autoReconnectPeriod;
     }
 
-    public BigDecimal getReadingPause() {
+    public int getReadingPause() {
         return readingPause;
     }
 
-    public BigDecimal getReadRetriesLimit() {
+    public int getReadRetriesLimit() {
         return readRetriesLimit;
     }
 
-    public BigDecimal getResponseTimeout() {
+    public int getResponseTimeout() {
         return responseTimeout;
     }
